Refactored the code into various folders. Made all the necessary changes. The map has been tested on multiple devices and it works.
