+++ app/src/main/java/edu/grinnell/appdev/events/Activities/SplashActivity.java
@@ -1,8 +1,10 @@
-package edu.grinnell.appdev.events;
+package edu.grinnell.appdev.events.Activities;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+
+import edu.grinnell.appdev.events.MainActivity;
 
 public class SplashActivity extends AppCompatActivity{
