+++ app/src/main/java/edu/grinnell/appdev/events/Parser/OnParseComplete.java
@@ -1,14 +1,16 @@
-package edu.grinnell.appdev.events;
+package edu.grinnell.appdev.events.Parser;
 
 
 import java.util.List;
+
+import edu.grinnell.appdev.events.Model.Event;
 
 
 /**
  * Provides methods to implement that is used by the Parser class
  * to notify the UI thread once the XML string has been parsed.
  */
-public interface onParseComplete {
+public interface OnParseComplete {
     void onParseComplete(List<Event> eventList);
     void onParseFail(String failMsg);
 }
